Refactor datastore config functions
- Separate datastore config functions into a separate file for each datastore
- Add .gitignore
